docs: relocate multi-GPU and single-subject acceleration benchmarks into Habitat Guide

+++ habit/habitat_features/each_habitat.py
@@ -211,7 +211,7 @@ class EachHabitatRadiomicsFeatures:
         as_: Optional[str] = None,
         use_torch_radiomics: Union[str, bool] = DEFAULT_USE_TORCH_RADIOMICS,
         torch_device: str = "auto",
-        torch_dtype: str = "float32",
+        torch_dtype: str = "float64",
     ) -> None:
         if modality is not None and modalities is not None:
             raise HABITAPIError(
